make the event_handler file easier to manage and add in what server the user joined in
